% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalWorstPower.R
\name{evalWorstPower}
\alias{evalWorstPower}
\title{Evaluate worst-case power}
\usage{
evalWorstPower(mt, freq, Amp, alpha = 0.05, method = c("eig", "test"))
}
\arguments{
\item{freq}{frequency of signal in same units as \code{t}}

\item{Amp}{amplitude of signal}

\item{alpha}{type I error rate, by default \code{alpha=0.05}}

\item{method}{by default, \code{method='eig'} uses eigenvalue method for computing the worstcase noncentrality parameter}

\item{t}{vector of measurement times in same units as \code{param$freq}}
}
\value{
minimum power
}
\description{
For a given frequency and amplitude, returns the lowest value of the power across all acrophases.
}
\note{
User can change method to \code{method='test'} to instead compute minimum power by discretising
the acrophase and searching for minimum. This will be slower and less accurate than the default so we
do not recommend using this unless the user is writing their own unit tests and wants to compare
\code{method='eig'} to an alternative method.
}
\examples{
# Worst-case power for a study with equispaced 24 timepoints
evalWorstPower(mt=1:24,param=list(Amp=1,freq=1/24))
}
\author{
Turner Silverthorne
}
