% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalWorstPowerMultiFreq.R
\name{evalWorstPowerMultiFreq}
\alias{evalWorstPowerMultiFreq}
\title{Evaluate worst-case power across multiple frequencies}
\usage{
evalWorstPowerMultiFreq(mt, param, alpha = 0.05, returnType = c("min", "all"))
}
\arguments{
\item{alpha}{type I error rate, by default \code{alpha=0.05}}

\item{returnType}{either \code{'min'} which returns the worst-case power across frequency and acrophase
or \code{'all'} which returns the worst case power of each frequency}

\item{t}{vector of measurement times assumed to lie in interval [0,1]}

\item{param$Amp}{amplitude of signal}

\item{param$fmin}{minimum frequency under consideration}

\item{param$fmax}{maximum frequency under consideration}
}
\value{
minimum power across acrophase and frequency range
}
\description{
For a given amplitude, return the lowest value of the power across all
acrophases and range of frequencies.
}
\author{
Turner Silverthorne
}
