% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exhaustiveSearch.R
\name{exhaustiveSearch}
\alias{exhaustiveSearch}
\title{Search all possible (discrete-time) designs to find optimal power}
\usage{
exhaustiveSearch(
  N,
  Nfine = 48,
  wlen,
  wdensity = 1,
  freq = 1,
  returnType = c("optimal", "all"),
  flines = 1e+06
)
}
\arguments{
\item{N}{number of measurements to be collected, aka sample size}

\item{Nfine}{number of points in the discrete grid}

\item{wlen}{length of the low-measurement window}

\item{wdensity}{number of measurements allowed in low-measurement window}

\item{freq}{frequency of signal}

\item{returnType}{if \code{returnType='optimal'} only the optimal solution is returned,
if \code{returnType='all'} then a dataframe containing one representative of each equivalence class
consistent with the timing constraint will be returned}

\item{flines}{number of lines to include in each awk output file.}
}
\value{
either the optimal solution consistent with the timing constraints, or a dataframe
containing all solutions consistent with timing constraints (see \code{returnType}).
}
\description{
There are only finitely many designs to consider if measurements are confined
to a discrete grid. Many designs will be equivalent up to temporal translation,
so only one representative of each equivalence class needs to be considered. This function
generates one representative from each equivalence class.
}
\note{
This function generates the database of designs using the [C library](http://combos.org/index)
and then filters this library using the \code{awk} script \code{c_src/necklace_filt.awk}.
}
\author{
Turner Silverthorne
}
