% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalExactPower.R
\name{evalExactPower}
\alias{evalExactPower}
\title{Evaluate exact power of cosinor model.}
\usage{
evalExactPower(
  t,
  param,
  alpha = 0.05,
  method = c("schur", "ncp", "equispaced"),
  lambda_in = NULL
)
}
\arguments{
\item{t}{vector of measurement times assumed to lie in interval [0,1]}

\item{alpha}{type I error, by default \code{alpha=.05}}

\item{method}{method for computing non-centrality parameter, by default
\code{method='schur'} uses the Schur complement formula.}

\item{lambda_in}{allows user to specify the noncentrality parameter instead of
calculating it from parameters of the signal.}

\item{param$Amp}{amplitude of signal}

\item{param$freq}{frequency of signal}

\item{param$acro}{phase of signal in radians}
}
\value{
Statistical power
}
\description{
Returns exact power of one-frequency harmonic regression hypothesis test
using the cosinor model.
}
\note{
Assumes the noise has mean zero and a standard deviation of one.

The default method of \code{method='schur'} should be sufficient for the vast
majority of use cases.   Alternative methods are as follows:
\code{'full'} which solves a linear system to obtain the noncentrality parameter,
\code{'ncp'} which lets user specify the noncentrality parameter, \code{'equispaced'} which uses
the power formula from \href{https://onlinelibrary.wiley.com/doi/full/10.1002/sim.9803}{Zong et al 2023},
which is only applicable for equispaced measurement times.
}
\examples{
# The power for a 24hr study sampled every hour, testing for p<0.05
evalExactPower(t=1:24,param=list(Amp=1,acro=0,freq=1/24))

}
\author{
Turner Silverthorne
}
