% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalMinEig.R
\name{evalMinEig}
\alias{evalMinEig}
\title{Evaluate smallest eigenvalue of power matrix.}
\usage{
evalMinEig(t, freq, design = c("general", "equispaced"))
}
\arguments{
\item{t}{vector of measurement times in same units as \code{freq}}

\item{freq}{frequency of signal in same units as \code{t}}

\item{design}{either \code{general} or \code{equispaced}}
}
\value{
minimum eigenvalue of the power matrix
}
\description{
Returns the smallest eigenvalue of the power matrix B^{-1} = (H (X^TX)^{-1} H^T)^{-1}. This eigenvalue is
equal to worst-case value of the noncentrality parameter across all signals with amplitude A=1
and noise strength sigma=1.
}
\note{
To calculate worst eigenvalue across multiple frequencies and acrophases,
use the wrapper function [evalMinEigMultiFreq()]
}
\author{
Turner Silverthorne
}
