% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalMinEigMultiFreq.R
\name{evalMinEigMultiFreq}
\alias{evalMinEigMultiFreq}
\title{Get worst eigenvalue across multiple frequencies}
\usage{
evalMinEigMultiFreq(mt, fmin, fmax, Nfreq, alpha = 0.05)
}
\arguments{
\item{fmin}{minium frequency, same units as \code{t}}

\item{fmax}{maximum frequency, same units as \code{t}}

\item{Nfreq}{number of points to use for discretising \code{[fmin,fmax]} interval}

\item{alpha}{type I error rate, by default \code{alpha=0.05}}

\item{t}{vector of measurement times, same units as \code{freq}}
}
\value{
minimum eigenvalue of the power matrix across the interval \code{[fmin,fmax]}
}
\description{
Wrapper for the function [evalMinEig()], computes worst eigenvalue of the power
matrix across multiple frequencies.
}
\author{
Turner Silverthorne
}
