% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalMonteCarloPower.R
\name{evalMonteCarloPower}
\alias{evalMonteCarloPower}
\title{Use Monte Carlo method to estimate power}
\usage{
evalMonteCarloPower(tvec, freq, acro, Amp, Nmc, alpha = 0.05)
}
\arguments{
\item{tvec}{vector of measurement times in same units as \code{param$freq}}

\item{freq}{frequency fo signal in same units as \code{t}}

\item{acro}{phase of signal in radians}

\item{Amp}{amplitude of signal}

\item{Nmc}{number of Monte Carlo samples}

\item{alpha}{type I error, default value \code{alpha=.05}}
}
\value{
Monte Carlo estimate of power
}
\description{
Performs harmonic regression on a simulated dataset of independent samples
and returns the portion of samples that have statistically significant p-values.
Useful for comparison with the exact expression for statistical power,
see [evalExactPower()].
}
\author{
Turner Silverthorne
}
