% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalWorstPower.R
\name{evalWorstPower}
\alias{evalWorstPower}
\title{Evaluate worst-case power}
\usage{
evalWorstPower(mt, param, alpha = 0.05, method = "eig")
}
\arguments{
\item{alpha}{type I error rate, by default \code{alpha=0.05}}

\item{method}{by default, \code{method='eig'} uses eigenvalue method for computing the worstcase noncentrality parameter}

\item{t}{vector of measurement times assumed to lie in interval [0,1]}

\item{param$Amp}{amplitude of signal}

\item{param$freq}{frequency of signal}
}
\value{
minimum power
}
\description{
For a given frequency and amplitude, returns the lowest value of the power across all acrophases.
}
\note{
User can change method to \code{method='test'} to instead compute minimum power by discretising
the acrophase and searching for minimum. This will be slower and less accurate than the default so we
do not recommend using this unless the user is writing their own unit tests and wants to compare
\code{method='eig'} to an alternative method.
}
\author{
Turner Silverthorne
}
